requirements.txt
django==4.1
django-extensions==3.2.0
djangorestframework==3.13.1
factory_boy==3.2.1
pre-commit==2.20.0
